% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_to_odd_window.R
\name{scale_to_odd_window}
\alias{scale_to_odd_window}
\title{Convert a target scale (map units) to an odd focal window size (cells)}
\usage{
scale_to_odd_window(r, scale_m, scale_type = c("diameter", "radius"))
}
\arguments{
\item{r}{A single-layer SpatRaster}

\item{scale_m}{Numeric. Spatial scale in metres.}

\item{scale_type}{Character. Interpretation of \code{scale}:
\itemize{
\item "diameter": window size ≈ scale / resolution
\item "radius":   window size ≈ (2 * scale) / resolution
}}
}
\value{
Integer odd window size (>= 1)
}
\description{
Convert a target scale (map units) to an odd focal window size (cells)
}
\keyword{internal}
