% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScalePredSDM-package.R
\docType{package}
\name{ScalePredSDM-package}
\alias{ScalePredSDM}
\alias{ScalePredSDM-package}
\title{ScalePredSDM: Scale-of-Effect Predictor Generation for Species Distribution Models}
\description{
Tools to generate, evaluate and assemble scale-of-effect environmental predictors for fine-scale species distribution models, with an emphasis on high-resolution marine and terrestrial data.
}
\author{
\strong{Maintainer}: Ole Sørensen \email{ole.johs.sorensen@gmail.com}

}
\keyword{internal}
